cmd/trace/bind: Move BindParser to a common package
