+++ cmd/kubectl-gadget/trace/bind.go
@@ -15,41 +15,18 @@
 package trace
 
 import (
-	"fmt"
 	"strconv"
 	"strings"
 
+	"github.com/spf13/cobra"
+
 	commontrace "github.com/kinvolk/inspektor-gadget/cmd/common/trace"
-	commonutils "github.com/kinvolk/inspektor-gadget/cmd/common/utils"
 	"github.com/kinvolk/inspektor-gadget/cmd/kubectl-gadget/utils"
 	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/trace/bind/types"
-	"github.com/spf13/cobra"
 )
 
-type BindParser struct {
-	commonutils.BaseParser[types.Event]
-}
-
 func newBindCmd() *cobra.Command {
-	commonFlags := &utils.CommonFlags{
-		OutputConfig: commonutils.OutputConfig{
-			// The columns that will be used in case the user does not specify
-			// which specific columns they want to print.
-			CustomColumns: []string{
-				"node",
-				"namespace",
-				"pod",
-				"container",
-				"pid",
-				"comm",
-				"proto",
-				"addr",
-				"port",
-				"opts",
-				"if",
-			},
-		},
-	}
+	var commonFlags utils.CommonFlags
 
 	var (
 		// flags
@@ -69,8 +46,8 @@ func newBindCmd() *cobra.Command {
 
 			bindGadget := &TraceGadget[types.Event]{
 				name:        "bindsnoop",
-				commonFlags: commonFlags,
-				parser:      NewBindParser(&commonFlags.OutputConfig),
+				commonFlags: &commonFlags,
+				parser:      commontrace.NewBindParserWithK8sInfo(&commonFlags.OutputConfig),
 				params: map[string]string{
 					"pid":           strconv.FormatUint(uint64(targetPid), 10),
 					"ports":         strings.Join(portsStringSlice, ","),
@@ -104,67 +81,7 @@ func newBindCmd() *cobra.Command {
 		"Show only events where the bind succeeded",
 	)
 
-	utils.AddCommonFlags(cmd, commonFlags)
+	utils.AddCommonFlags(cmd, &commonFlags)
 
 	return cmd
-}
-
-func NewBindParser(outputConfig *commonutils.OutputConfig) commontrace.TraceParser[types.Event] {
-	columnsWidth := map[string]int{
-		"node":      -16,
-		"namespace": -16,
-		"pod":       -30,
-		"container": -16,
-		"pid":       -7,
-		"comm":      -16,
-		"proto":     -6,
-		"addr":      -16,
-		"port":      -7,
-		"opts":      -7,
-		"if":        -7,
-	}
-
-	return &BindParser{
-		BaseParser: commonutils.NewBaseWidthParser[types.Event](columnsWidth, outputConfig),
-	}
-}
-
-func (p *BindParser) TransformEvent(event *types.Event) string {
-	return p.Transform(event, func(event *types.Event) string {
-		var sb strings.Builder
-
-		for _, col := range p.OutputConfig.CustomColumns {
-			switch col {
-			case "node":
-				sb.WriteString(fmt.Sprintf("%*s", p.ColumnsWidth[col], event.Node))
-			case "namespace":
-				sb.WriteString(fmt.Sprintf("%*s", p.ColumnsWidth[col], event.Namespace))
-			case "pod":
-				sb.WriteString(fmt.Sprintf("%*s", p.ColumnsWidth[col], event.Pod))
-			case "container":
-				sb.WriteString(fmt.Sprintf("%*s", p.ColumnsWidth[col], event.Container))
-			case "pid":
-				sb.WriteString(fmt.Sprintf("%*d", p.ColumnsWidth[col], event.Pid))
-			case "comm":
-				sb.WriteString(fmt.Sprintf("%*s", p.ColumnsWidth[col], event.Comm))
-			case "proto":
-				sb.WriteString(fmt.Sprintf("%*s", p.ColumnsWidth[col], event.Protocol))
-			case "addr":
-				sb.WriteString(fmt.Sprintf("%*s", p.ColumnsWidth[col], event.Addr))
-			case "port":
-				sb.WriteString(fmt.Sprintf("%*d", p.ColumnsWidth[col], event.Port))
-			case "opts":
-				sb.WriteString(fmt.Sprintf("%*s", p.ColumnsWidth[col], event.Options))
-			case "if":
-				sb.WriteString(fmt.Sprintf("%*s", p.ColumnsWidth[col], event.Interface))
-			default:
-				continue
-			}
-
-			// Needed when field is larger than the predefined columnsWidth.
-			sb.WriteRune(' ')
-		}
-
-		return sb.String()
-	})
 }
